figured out what was wrong with $au_mslp_temp
so NCEP changes the one of the requested variables of the sounding every run. so midnight = 0, 06z = 1, 12z = 2, and 18z = 3. 

I figured this out after spending an entire day wondering who was messing with my script... how silly. anyways, i just created a function to download the .nc file to my computer and create it from there.