updates to sfc charts to fix NCEP variable issue
the problem with the mslp_temp surface charts was due to the fact that
NCEP changes the 'reftime' variable every sounding run change from
'reftime' for 00z to 'reftime1' for 06z to 'reftime2' for 12z
and finally 'reftime3' for 18z run. For some reason this only applies
to the isobaric MSLP charts. all sfc maps have been fixed and future
ones will not have any issues